backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
openai

# Environment variables
python-dotenv==1.0.0

# Audio processing
pydub==0.25.1
numpy==1.24.3

# Additional dependencies
requests==2.30.0